Update 1234.txt
In 1998, before the company was even incorporated, the concept of the doodle was born when Google founders Larry and Sergey played with the corporate logo to indicate their attendance at the Burning Man festival in the Nevada desert. They placed a stick figure drawing behind the 2nd "o" in the word, Google, and the revised logo was intended as a comical message to Google users that the founders were "out of office.". While the first doodle was relatively

the concept of the doodle was born when Google founders Larry and Sergey played with the corporate logo to indicate their attendance at the Burning Man festival in the Nevada desert. They placed a stick figure drawing behind the 2nd "o" in the word, Google, and the revised logo was intended as a comical message to Google users that the founders were "out of office.". While the first doodle was relatively simple, the idea of decorating the company logo to celebrate notable events was born.